+++ src/rubicon/objc/api.py
@@ -158,16 +158,18 @@ class ObjCMethod:
 
     :class:`ObjCMethod` objects normally don't need to be used directly. To call
     a method on an Objective-C object, you should use the method call syntax
-    supported by :class:`ObjCInstance`, or the
+    supported by [`ObjCInstance`][rubicon.objc.api.ObjCInstance], or the
     :func:`~rubicon.objc.runtime.send_message` function.
 
-    .. note::
+    /// note | Note
 
-        This is *not* the same class as the one used for *bound* Objective-C
-        methods, as returned from :meth:`ObjCInstance.__getattr__`. Currently,
-        Rubicon doesn't provide any documented way to get an unbound
-        :class:`ObjCMethod` object for an instance method of an
-        :class:`ObjCClass`.
+    This is *not* the same class as the one used for *bound* Objective-C
+    methods, as returned from :meth:`ObjCInstance.__getattr__`. Currently,
+    Rubicon doesn't provide any documented way to get an unbound
+    :class:`ObjCMethod` object for an instance method of an
+    [`ObjCClass`][rubicon.objc.api.ObjCClass].
+
+    ///
     """
 
     def __init__(self, method):
@@ -200,7 +202,7 @@ class ObjCMethod:
         * :class:`enum.Enum` objects are replaced by their
           :attr:`~enum.Enum.value` before further conversion
         * For parameters that expect a block, Python callables are converted to
-          :class:`Block`\\s
+          :class:`Block`s
         * For parameters that expect an Objective-C object, Python objects are
           converted using :func:`ns_from_py`
         * For parameters that expect a C structure, Python sequences are
@@ -214,7 +216,8 @@ class ObjCMethod:
         ``False``.
 
         If the method returns an Objective-C object, it is automatically
-        converted to an :class:`ObjCInstance`. This conversion can be disabled
+        converted to an [`ObjCInstance`][rubicon.objc.api.ObjCInstance]. This
+        conversion can be disabled
         by setting the ``convert_result`` keyword argument to ``False``, in
         which case the object is returned as a raw
         :class:`~rubicon.objc.runtime.objc_id` value.
@@ -717,13 +720,14 @@ _type_for_objcclass_map = {}
 
 
 def type_for_objcclass(objcclass):
-    """Look up the :class:`ObjCInstance` subclass used to represent instances of the
-    given Objective-C class in Python.
+    """Look up the [`ObjCInstance`][rubicon.objc.api.ObjCInstance] subclass used to
+    represent instances of the given Objective-C class in Python.
 
     If the exact Objective-C class is not registered, each superclass is also
-    checked, defaulting to :class:`ObjCInstance` if none of the classes in the
-    superclass chain is registered. Afterwards, all searched superclasses are
-    registered for the :class:`ObjCInstance` subclass that was found. (This
+    checked, defaulting to [`ObjCInstance`][rubicon.objc.api.ObjCInstance] if none of
+    the classes in the superclass chain is registered. Afterward, all searched
+    superclasses are registered for the [`ObjCInstance`][rubicon.objc.api.ObjCInstance]
+    subclass that was found. (This
     speeds up future lookups, and ensures that previously computed mappings are
     not changed by unrelated registrations.)
 
@@ -753,21 +757,23 @@ def type_for_objcclass(objcclass):
 
 
 def register_type_for_objcclass(pytype, objcclass):
-    """Register a conversion from an Objective-C class to an :class:`ObjCInstance`
-    subclass.
+    """Register a conversion from an Objective-C class to an
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] subclass.
 
     After a call of this function, when Rubicon wraps an Objective-C object that
     is an instance of ``objcclass`` (or a subclass), the Python object will have
-    the class ``pytype`` rather than :class:`ObjCInstance`. See
-    :func:`type_for_objcclass` for a full description of the lookup process.
+    the class ``pytype`` rather than [`ObjCInstance`][rubicon.objc.api.ObjCInstance].
+    See :func:`type_for_objcclass` for a full description of the lookup process.
 
-    .. warning::
+    /// warning | Warning
 
-        This function should only be called if no instances of ``objcclass`` (or
-        a subclass) have been wrapped by Rubicon yet. If the function is called
-        later, it will not fully take effect: the types of existing instances do
-        not change, and mappings for subclasses of ``objcclass`` are not
-        updated.
+    This function should only be called if no instances of ``objcclass`` (or
+    a subclass) have been wrapped by Rubicon yet. If the function is called
+    later, it will not fully take effect: the types of existing instances do
+    not change, and mappings for subclasses of ``objcclass`` are not
+    updated.
+
+    ///
     """
 
     if isinstance(objcclass, ObjCClass):
@@ -777,16 +783,18 @@ def register_type_for_objcclass(pytype, objcclass):
 
 
 def unregister_type_for_objcclass(objcclass):
-    """Unregister a conversion from an Objective-C class to an :class:`ObjCInstance`
-    subclass.
+    """Unregister a conversion from an Objective-C class to an
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] subclass.
 
-    .. warning::
+    /// warning | Warning
 
-        This function should only be called if no instances of ``objcclass`` (or
-        a subclass) have been wrapped by Rubicon yet. If the function is called
-        later, it will not fully take effect: the types of existing instances do
-        not change, and mappings for subclasses of ``objcclass`` are not
-        removed.
+    This function should only be called if no instances of ``objcclass`` (or
+    a subclass) have been wrapped by Rubicon yet. If the function is called
+    later, it will not fully take effect: the types of existing instances do
+    not change, and mappings for subclasses of ``objcclass`` are not
+    removed.
+
+    ///
     """
 
     if isinstance(objcclass, ObjCClass):
@@ -796,10 +804,10 @@ def unregister_type_for_objcclass(objcclass):
 
 
 def get_type_for_objcclass_map():
-    """Get a copy of all currently registered :class:`ObjCInstance` subclasses as a
-    mapping.
+    """Get a copy of all currently registered
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] subclasses as a mapping.
 
-    Keys are Objective-C class addresses as :class:`int`\\s.
+    Keys are Objective-C class addresses as :class:`int`s.
     """
 
     return dict(_type_for_objcclass_map)
@@ -807,7 +815,7 @@ def get_type_for_objcclass_map():
 
 def for_objcclass(objcclass):
     """Decorator for registering a conversion from an Objective-C class to an
-    :class:`ObjCInstance` subclass.
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] subclass.
 
     This is equivalent to calling :func:`register_type_for_objcclass` on
     the decorated class.
@@ -821,20 +829,53 @@ def for_objcclass(objcclass):
 
 
 class ObjCInstance:
-    """Python wrapper for an Objective-C instance."""
+    """Python wrapper for an Objective-C instance.
+
+    The constructor accepts an :class:`~rubicon.objc.runtime.objc_id` or anything
+    that can be cast to one, such as a :class:`~ctypes.c_void_p`, or an existing
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance].
+
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] objects are cached ---
+    this means that for every
+    Objective-C object there can be at most on
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] object
+    at any time. Rubicon will automatically create new
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance]s or return existing ones as needed.
+
+    The returned object's Python class is not always exactly
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance]. For example, if the passed
+    pointer refers to a class or a metaclass, an instance o
+    [`ObjCClass`][rubicon.objc.api.ObjCClass] or
+    [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] is returned as appropriate.
+    Additional custom
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] subclasses may be defined and
+    registered using
+    :func:`register_type_for_objcclass`. Creating an
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance]
+    from a ``nil`` pointer returns ``None``.
+
+    Rubicon retains an Objective-C object when it is wrapped in an
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] and autoreleases it when the
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] is
+    garbage collected.
+
+    The only exception to this are objects returned by methods which create an
+    object (starting with "alloc", "new", "copy", or "mutableCopy"). We do not
+    explicitly retain them because we already own objects created by us, but we do
+    autorelease them on garbage collection of the Python wrapper.
+
+    This ensures that the [`ObjCInstance`][rubicon.objc.api.ObjCInstance] can always be
+    used from Python
+    without segfaults while preventing Rubicon from leaking memory.
+    """
 
     ptr: objc_id
     """The wrapped object pointer as an [`objc_id`][rubicon.objc.objc_id].
 
-    This
-    attribute is also available as `_as_parameter_` to allow
+    This attribute is also available as `_as_parameter_` to allow
     [`ObjCInstance`][rubicon.objc.api.ObjCInstance]s to be passed into
     [`ctypes`][] functions.
     """
-    _as_parameter_: objc_id
-    """[`ptr`][rubicon.objc.api.ObjCInstance.ptr] available as `_as_parameter_` to allow
-    [`ObjCInstance`][rubicon.objc.api.ObjCInstance]s to be passed into [`ctypes`][]
-    functions."""
 
     # Cache dictionary containing every currently existing ObjCInstance object,
     # with the key being the memory address (as an integer) of the Objective-C
@@ -861,7 +902,8 @@ class ObjCInstance:
 
     @property
     def objc_class(self):
-        """The Objective-C object's class, as an :class:`ObjCClass`."""
+        """The Objective-C object's class, as an
+        [`ObjCClass`][rubicon.objc.api.ObjCClass]."""
 
         # This property is used inside __getattr__, so any attribute accesses must be
         # done through super(...).__getattribute__ to prevent infinite recursion.
@@ -888,35 +930,44 @@ class ObjCInstance:
     def __new__(
         cls, object_ptr, _name=None, _bases=None, _ns=None, _implicitly_owned=False
     ):
-        """The constructor accepts an :class:`~rubicon.objc.runtime.objc_id` or anything
-        that can be cast to one, such as a :class:`~ctypes.c_void_p`, or an existing
-        :class:`ObjCInstance`.
-
-        :class:`ObjCInstance` objects are cached --- this means that for every
-        Objective-C object there can be at most one :class:`ObjCInstance` object
-        at any time. Rubicon will automatically create new
-        :class:`ObjCInstance`\\s or return existing ones as needed.
-
-        The returned object's Python class is not always exactly
-        :class:`ObjCInstance`. For example, if the passed pointer refers to a
-        class or a metaclass, an instance of :class:`ObjCClass` or
-        :class:`ObjCMetaClass` is returned as appropriate. Additional custom
-        :class:`ObjCInstance` subclasses may be defined and registered using
-        :func:`register_type_for_objcclass`. Creating an :class:`ObjCInstance`
-        from a ``nil`` pointer returns ``None``.
-
-        Rubicon retains an Objective-C object when it is wrapped in an
-        :class:`ObjCInstance` and autoreleases it when the :class:`ObjCInstance` is
-        garbage collected.
-
-        The only exception to this are objects returned by methods which create an
-        object (starting with "alloc", "new", "copy", or "mutableCopy"). We do not
-        explicitly retain them because we already own objects created by us, but we do
-        autorelease them on garbage collection of the Python wrapper.
-
-        This ensures that the :class:`ObjCInstance` can always be used from Python
-        without segfaults while preventing Rubicon from leaking memory.
-        """
+        # The constructor accepts an :class:`~rubicon.objc.runtime.objc_id` or anything
+        # that can be cast to one, such as a :class:`~ctypes.c_void_p`, or an existing
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance].
+        #
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] objects are cached --- this
+        # means that for every
+        # Objective-C object there can be at most one
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] object
+        # at any time. Rubicon will automatically create new
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance]s or return existing ones as
+        # needed.
+        #
+        # The returned object's Python class is not always exactly
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance]. For example, if the passed
+        # pointer refers to a
+        # class or a metaclass, an instance of [`ObjCClass`][rubicon.objc.api.ObjCClass]
+        # or
+        # [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] is returned as appropriate.
+        # Additional custom
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] subclasses may be defined and
+        # registered using
+        # :func:`register_type_for_objcclass`. Creating an
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance]
+        # from a ``nil`` pointer returns ``None``.
+        #
+        # Rubicon retains an Objective-C object when it is wrapped in an
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] and autoreleases it when the
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] is
+        # garbage collected.
+        #
+        # The only exception to this are objects returned by methods which create an
+        # object (starting with "alloc", "new", "copy", or "mutableCopy"). We do not
+        # explicitly retain them because we already own objects created by us, but we do
+        # autorelease them on garbage collection of the Python wrapper.
+        #
+        # This ensures that the [`ObjCInstance`][rubicon.objc.api.ObjCInstance] can
+        # always be used from Python
+        # without segfaults while preventing Rubicon from leaking memory.
 
         # Make sure that object_ptr is wrapped in an objc_id.
         if not isinstance(object_ptr, objc_id):
@@ -1043,8 +1094,8 @@ class ObjCInstance:
           :meth:`ObjCClass.declare_property`
 
         Otherwise, a method matching the given name is looked up.
-        :class:`ObjCInstance` understands two syntaxes for calling Objective-C
-        methods:
+        [`ObjCInstance`][rubicon.objc.api.ObjCInstance] understands two syntaxes for
+        calling Objective-C methods:
 
         * "Flat" syntax: the Objective-C method name is spelled out in the
           attribute name, with all colons replaced with underscores, and all
@@ -1193,10 +1244,41 @@ class ObjCInstance:
 class ObjCClass(ObjCInstance, type):
     """Python wrapper for an Objective-C class.
 
-    :class:`ObjCClass` is a subclass of :class:`ObjCInstance` and supports the
+    [`ObjCClass`][rubicon.objc.api.ObjCClass] is a subclass of
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] and supports the
     same syntaxes for calling methods and accessing properties.
 
-    TODO: Add name.
+    The constructor accepts either the name of an Objective-C class to look up
+    (as [`str`][] or [`bytes`][]), or a pointer to an existing class object
+    (in any form accepted by [`ObjCInstance`][rubicon.objc.api.ObjCInstance]).
+
+    If given a pointer, it must refer to an Objective-C class; pointers to
+    other objects are not accepted. (Use
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] to wrap a
+    pointer that might also refer to other kinds of objects.) If the pointer
+    refers to a metaclass, an instance of
+    [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] is returned
+    instead. Creating an [`ObjCClass`][rubicon.objc.api.ObjCClass] from a ``Nil``
+    pointer returns ``None``.
+
+    [`ObjCClass`][rubicon.objc.api.ObjCClass] can also be called like
+    :class:`type`, with three
+    arguments (name, bases list, namespace mapping). This form is called
+    implicitly by Python's ``class`` syntax, and is used to create a new
+    Objective-C class from Python (see :ref:`custom-classes-and-protocols`).
+    The bases list must contain exactly one
+    [`ObjCClass`][rubicon.objc.api.ObjCClass] to be
+    extended by the new class. An optional ``protocols`` keyword argument is
+    also accepted, which must be a sequence of :class:`ObjCProtocol`s for
+    the new class to adopt.
+
+    If the name of the class has already registered with the Objective-C
+    runtime, the ``auto_rename`` option can be used to ensure that the
+    Objective-C name for the new class will be unique. A numeric suffix will
+    be appended to the Objective-C name to ensure uniqueness (for example,
+    ``MyClass`` will be renamed to ``MyClass_2``, ``MyClass_3`` etc until a
+    unique name is found). By default, classes will *not* be renamed, unless
+    :attr:`ObjCClass.auto_rename` is set at the class level.
     """
 
     name: str
@@ -1225,7 +1307,7 @@ class ObjCClass(ObjCInstance, type):
     """
     A :any:`bool` value describing whether a defined class should be renamed
     automatically if a class with the same name already exists in the
-    Objective C runtime.
+    Objective-C runtime.
     """
 
     @classmethod
@@ -1358,34 +1440,37 @@ class ObjCClass(ObjCInstance, type):
         protocols=(),
         auto_rename=None,
     ):
-        """The constructor accepts either the name of an Objective-C class to look up
-        (as :class:`str` or :class:`bytes`), or a pointer to an existing class object
-        (in any form accepted by :class:`ObjCInstance`).
-
-        If given a pointer, it must refer to an Objective-C class; pointers to
-        other objects are not accepted. (Use :class:`ObjCInstance` to wrap a
-        pointer that might also refer to other kinds of objects.) If the pointer
-        refers to a metaclass, an instance of :class:`ObjCMetaClass` is returned
-        instead. Creating an :class:`ObjCClass` from a ``Nil`` pointer returns
-        ``None``.
-
-        :class:`ObjCClass` can also be called like :class:`type`, with three
-        arguments (name, bases list, namespace mapping). This form is called
-        implicitly by Python's ``class`` syntax, and is used to create a new
-        Objective-C class from Python (see :ref:`custom-classes-and-protocols`).
-        The bases list must contain exactly one :class:`ObjCClass` to be
-        extended by the new class. An optional ``protocols`` keyword argument is
-        also accepted, which must be a sequence of :class:`ObjCProtocol`\\s for
-        the new class to adopt.
-
-        If the name of the class has already registered with the Objective C
-        runtime, the ``auto_rename`` option can be used to ensure that the
-        Objective C name for the new class will be unique. A numeric suffix will
-        be appended to the Objective C name to ensure uniqueness (for example,
-        ``MyClass`` will be renamed to ``MyClass_2``, ``MyClass_3`` etc until a
-        unique name is found). By default, classes will *not* be renamed, unless
-        :attr:`ObjCClass.auto_rename` is set at the class level.
-        """
+        # The constructor accepts either the name of an Objective-C class to look up
+        # (as [`str`][] or [`bytes`][]), or a pointer to an existing class object
+        # (in any form accepted by [`ObjCInstance`][rubicon.objc.api.ObjCInstance]).
+        #
+        # If given a pointer, it must refer to an Objective-C class; pointers to
+        # other objects are not accepted. (Use
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] to wrap a
+        # pointer that might also refer to other kinds of objects.) If the pointer
+        # refers to a metaclass, an instance of
+        # [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] is returned
+        # instead. Creating an [`ObjCClass`][rubicon.objc.api.ObjCClass] from a ``Nil``
+        # pointer returns ``None``.
+        #
+        # [`ObjCClass`][rubicon.objc.api.ObjCClass] can also be called like
+        # :class:`type`, with three
+        # arguments (name, bases list, namespace mapping). This form is called
+        # implicitly by Python's ``class`` syntax, and is used to create a new
+        # Objective-C class from Python (see :ref:`custom-classes-and-protocols`).
+        # The bases list must contain exactly one
+        # [`ObjCClass`][rubicon.objc.api.ObjCClass] to be
+        # extended by the new class. An optional ``protocols`` keyword argument is
+        # also accepted, which must be a sequence of :class:`ObjCProtocol`s for
+        # the new class to adopt.
+        #
+        # If the name of the class has already registered with the Objective-C
+        # runtime, the ``auto_rename`` option can be used to ensure that the
+        # Objective-C name for the new class will be unique. A numeric suffix will
+        # be appended to the Objective-C name to ensure uniqueness (for example,
+        # ``MyClass`` will be renamed to ``MyClass_2``, ``MyClass_3`` etc until a
+        # unique name is found). By default, classes will *not* be renamed, unless
+        # :attr:`ObjCClass.auto_rename` is set at the class level.
 
         if (bases is None) ^ (attrs is None):
             raise TypeError("ObjCClass arguments 2 and 3 must be given together")
@@ -1578,22 +1663,24 @@ class ObjCClass(ObjCInstance, type):
         ``@property`` declaration in the source code, because Rubicon cannot tell such
         properties apart from normal zero-argument methods.
 
-        .. note::
+        /// note | Note
 
-            In the standard Apple SDKs, some properties are introduced as regular
-            methods in one system version, and then declared as properties in a later
-            system version. For example, the ``description`` method/property of
-            :class:`NSObject` was declared as a regular method `up to OS X 10.9
-            <https://github.com/phracker/MacOSX-SDKs/blob/9fc3ed0ad0345950ac25c28695b0427846eea966/MacOSX10.9.sdk/usr/include/objc/NSObject.h#L40>`__,
-            but changed to a property `as of OS X 10.10
-            <https://github.com/phracker/MacOSX-SDKs/blob/9fc3ed0ad0345950ac25c28695b0427846eea966/MacOSX10.10.sdk/usr/include/objc/NSObject.h#L43>`__.
+        In the standard Apple SDKs, some properties are introduced as regular
+        methods in one system version, and then declared as properties in a later
+        system version. For example, the ``description`` method/property of
+        :class:`NSObject` was declared as a regular method `up to OS X 10.9
+        <https://github.com/phracker/MacOSX-SDKs/blob/9fc3ed0ad0345950ac25c28695b0427846eea966/MacOSX10.9.sdk/usr/include/objc/NSObject.h#L40>`__,
+        but changed to a property `as of OS X 10.10
+        <https://github.com/phracker/MacOSX-SDKs/blob/9fc3ed0ad0345950ac25c28695b0427846eea966/MacOSX10.10.sdk/usr/include/objc/NSObject.h#L43>`__.
 
-            Such properties cause compatibility issues when accessed from Rubicon:
-            ``obj.description()`` works on 10.9 but is a :class:`TypeError` on 10.10,
-            whereas ``obj.description`` works on 10.10 but returns a method object on
-            10.9. To solve this issue, the property can be declared explicitly using
-            ``NSObject.declare_property('description')``, so that it can always be
-            accessed using ``obj.description``.
+        Such properties cause compatibility issues when accessed from Rubicon:
+        ``obj.description()`` works on 10.9 but is a [`TypeError`][] on 10.10,
+        whereas ``obj.description`` works on 10.10 but returns a method object on
+        10.9. To solve this issue, the property can be declared explicitly using
+        ``NSObject.declare_property('description')``, so that it can always be
+        accessed using ``obj.description``.
+
+        ///
         """
 
         self.forced_properties.add(name)
@@ -1621,7 +1708,8 @@ class ObjCClass(ObjCInstance, type):
 
         If the given object is not an Objective-C object, ``False`` is returned.
 
-        This method allows using :class:`ObjCClass`\\es as the second argument
+        This method allows using [`ObjCClass`][rubicon.objc.api.ObjCClass]es as the
+        second argument
         of :func:`isinstance`: ``isinstance(obj, NSString)`` is equivalent to
         ``obj.isKindOfClass(NSString)``.
         """
@@ -1634,10 +1722,11 @@ class ObjCClass(ObjCInstance, type):
     def __subclasscheck__(self, subclass):
         """Check whether the given class is a subclass of this class.
 
-        If the given object is not an Objective-C class, :class:`TypeError` is
+        If the given object is not an Objective-C class, [`TypeError`][] is
         raised.
 
-        This method allows using :class:`ObjCClass`\\es as the second argument
+        This method allows using [`ObjCClass`][rubicon.objc.api.ObjCClass]es as the
+        second argument
         of :func:`issubclass`: ``issubclass(cls, NSValue)`` is equivalent to
         ``obj.isSubclassOfClass(NSValue)``.
         """
@@ -1696,25 +1785,45 @@ class ObjCClass(ObjCInstance, type):
 class ObjCMetaClass(ObjCClass):
     """Python wrapper for an Objective-C metaclass.
 
-    :class:`ObjCMetaClass` is a subclass of :class:`ObjCClass` and supports
+    [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] is a subclass of
+    [`ObjCClass`][rubicon.objc.api.ObjCClass] and supports
     almost exactly the same operations and methods. However, there is usually no
     need to look up a metaclass manually. The main reason why
-    :class:`ObjCMetaClass` is a separate class is to differentiate it from
-    :class:`ObjCClass` in the :func:`repr`. (Otherwise there would be no way to
+    [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] is a separate class is to
+    differentiate it from
+    [`ObjCClass`][rubicon.objc.api.ObjCClass] in the [`repr`][]. (Otherwise there
+    would be no way to
     tell classes and metaclasses apart, since metaclasses are also classes, and
     have exactly the same name as their corresponding class.)
+
+    The constructor accepts either the name of an Objective-C metaclass to look
+    up (as [`str`][] or [`bytes`][]), or a pointer to an existing metaclass
+    object (in any form accepted by
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance]).
+
+    If given a pointer, it must refer to an Objective-C metaclass; pointers
+    to other objects are not accepted. (Use
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] to wrap a
+    pointer that might also refer to other kinds of objects.) Creating an
+    [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] from a ``Nil`` pointer
+    returns ``None``.
     """
 
-    def __new__(cls, name_or_ptr):
-        """The constructor accepts either the name of an Objective-C metaclass to look
-        up (as :class:`str` or :class:`bytes`), or a pointer to an existing metaclass
-        object (in any form accepted by :class:`ObjCInstance`).
+    name: str
+    """The name of this class, as a [`str`][]."""
 
-        If given a pointer, it must refer to an Objective-C metaclass; pointers
-        to other objects are not accepted. (Use :class:`ObjCInstance` to wrap a
-        pointer that might also refer to other kinds of objects.) Creating an
-        :class:`ObjCMetaClass` from a ``Nil`` pointer returns ``None``.
-        """
+    def __new__(cls, name_or_ptr):
+        # The constructor accepts either the name of an Objective-C metaclass to look
+        # up (as [`str`][] or [`bytes`][]), or a pointer to an existing metaclass
+        # object (in any form accepted by
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance]).
+        #
+        # If given a pointer, it must refer to an Objective-C metaclass; pointers
+        # to other objects are not accepted. (Use
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] to wrap a
+        # pointer that might also refer to other kinds of objects.) Creating an
+        # [`ObjCMetaClass`][rubicon.objc.api.ObjCMetaClass] from a ``Nil`` pointer
+        # returns ``None``.
 
         if isinstance(name_or_ptr, (bytes, str)):
             name = ensure_bytes(name_or_ptr)
@@ -1748,15 +1857,25 @@ if True:
     See the [`ObjCInstance`][rubicon.objc.api.ObjCInstance]
     documentation for a list of operations that Rubicon supports on all objects.
 
-    # `debugDescription`
-    This Objective-C property has been declared using
-    [`ObjCClass.declare_property()`][rubicon.objc.api.ObjCClass.declare_property] and
-    can always be accessed using attribute syntax.
+    # debugDescription
 
-    # `description`
-    This Objective-C property has been declared using
-    [`ObjCClass.declare_property()`][rubicon.objc.api.ObjCClass.declare_property] and
-    can always be accessed using attribute syntax.
+    ```
+    debugDescription
+    ```
+
+    Exposes the Objective-C
+    [`debugDescription`](https://developer.apple.com/documentation/objectivec/nsobjectprotocol/debugdescription?language=objc)
+    property.
+
+    # description
+
+    ```
+    description
+    ```
+
+    Exposes the Objective-C
+    [`description`](https://developer.apple.com/documentation/objectivec/nsobjectprotocol/description?language=objc)
+    property.
     """
     NSObject.declare_property("debugDescription")
     NSObject.declare_property("description")
@@ -1766,13 +1885,9 @@ if True:
     [NSNumber](https://developer.apple.com/documentation/foundation/nsnumber?language=objc)
     class from `<Foundation/NSValue.h>`.
 
-    /// note | Note
-
     This class can be converted to and from standard Python primitives
     (`bool`, `int`, `float`) using [`py_from_ns`][rubicon.objc.py_from_ns] and
     [`ns_from_py`][rubicon.objc.ns_from_py].
-
-    ///
     """
     NSDecimalNumber = ObjCClass("NSDecimalNumber")
     """
@@ -1780,23 +1895,17 @@ if True:
     [NSDecimalNumber](https://developer.apple.com/documentation/foundation/nsdecimalnumber?language=objc)
     class from `<Foundation/NSDecimalNumber.h>`.
 
-    /// note | Note
-
     This class can be converted to and from Python `decimal.Decimal` using
     [`py_from_ns`][rubicon.objc.py_from_ns] and
     [`ns_from_py`][rubicon.objc.ns_from_py].
-
-    ///
     """
     NSString = ObjCClass("NSString")
-    """
+    r"""
     The
     [NSString](https://developer.apple.com/documentation/foundation/nsstring?language=objc)
     class from `<Foundation/NSString.h>`.
 
     This class also supports all methods that [`str`][] does.
-
-    /// note | Note
 
     This class can be converted to and from Python [`str`][] using
     [`py_from_ns`][rubicon.objc.py_from_ns] and
@@ -1811,13 +1920,21 @@ if True:
     points is required, convert the [`NSString`][rubicon.objc.api.NSString] to
     [`str`][] first.
 
-    ///
+    # \_\_str\_\_
 
-    # `str`
+    ```python
+    __str__()
+    ```
+
     Return the value of this [`NSString`][rubicon.objc.api.NSString] as a
     [`str`][].
 
-    # `UTF8String`
+    # UTF8String
+
+    ```python
+    UTF8String
+    ```
+
     This Objective-C property has been declared using
     [`ObjCClass.declare_property`][rubicon.objc.api.ObjCClass.declare_property]
     and can always be accessed using attribute syntax.
@@ -1829,21 +1946,15 @@ if True:
     [NSData](https://developer.apple.com/documentation/foundation/nsdata?language=objc)
     class from `<Foundation/NSData.h>`.
 
-    /// note | Note
-
     This class can be converted to and from Python [`bytes`][] using
     [`py_from_ns`][rubicon.objc.py_from_ns] and
     [`ns_from_py`][rubicon.objc.ns_from_py].
-
-    ///
     """
     NSArray = ObjCClass("NSArray")
     """
     The
     [NSArray](https://developer.apple.com/documentation/foundation/nsarray?language=objc)
     class from `<Foundation/NSArray.h>`.
-
-    /// note | Note
 
     This class can be converted to and from Python [`list`][] using
     [`py_from_ns`][rubicon.objc.py_from_ns] and
@@ -1859,18 +1970,10 @@ if True:
     any of `pylist`'s elements cannot be converted to Objective-C, an error
     is raised.
 
-    ///
-
-    # `__getitem__(index)`
-    # `__len__()`
-    # `__iter__()`
-    # `__contains__(value)`
-    # `__eq__(other)`
-    # `__ne__(other)`
-    # `index(value)`
-    # `count(value)`
-    # `copy()`
-    Python-style sequence interface.
+    Supports
+    [Python-style sequence operations](https://docs.python.org/3/library/stdtypes.html#typesseq)
+    including: `__getitem__()`, `__len__()`, `__iter__()`, `__contains__()`, `__eq__()`,
+    `__ne__()`, `index()`, `count()`, and `copy()`.
     """
     NSMutableArray = ObjCClass("NSMutableArray")
     """
@@ -1878,41 +1981,93 @@ if True:
     [NSMutableArray](https://developer.apple.com/documentation/foundation/nsmutablearray?language=objc)
     class from `<Foundation/NSArray.h>`.
 
-    /// note | Note
-
     This class can be converted to and from Python exactly like its
     superclass `NSArray`.
 
-    ///
-
-    # `__setitem__(index, value)`
-    # `__delitem__(index)`
-    # `append(value)`
-    # `clear()`
-    # `extend(values)`
-    # `insert(index, value)`
-    # `pop([index=-1])`
-    # `remove(value)`
-    # `reverse()`
-    Python-style mutable sequence interface.
+    Supports
+    [Python-style mutable sequence operations](https://docs.python.org/3/library/stdtypes.html#mutable-sequence-types)
+    including: `__setitem__()`, `__delitem__()`, `append()`, `clear()`, `extend()`,
+    `insert()`, `pop()`, `remove()`, and `reverse()`.
     """
     NSDictionary = ObjCClass("NSDictionary")
+    """
+    The
+    [NSDictionary](https://developer.apple.com/documentation/foundation/nsdictionary?language=objc)
+    class from `<Foundation/NSDictionary.h>`.
+
+    This class can be converted to and from Python [`dict`][] using
+    [`py_from_ns`][rubicon.objc.py_from_ns] and
+    [`ns_from_py`][rubicon.objc.ns_from_py].
+
+    `py_from_ns(nsdict)` will recursively convert `nsdict`'s keys and values
+    to Python objects, where possible. To avoid the recursive conversion of
+    the values, use `{py_from_ns(k): v for k, v in nsdict.items()}`. The
+    conversion of the keys cannot be avoided, because Python
+    [`dict][] keys need to be hashable, which
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] is not. If any of the
+    keys convert to a Python object that is not hashable, an error is raised
+    (regardless of which conversion method you use).
+
+    `ns_from_py(pydict)` will recursively convert `pydict`'s keys and values
+    to Objective-C. As there is no way to store Python object references as
+    Objective-C objects yet, this recursive conversion cannot be avoided. If
+    any of `pydict`'s keys or values cannot be converted to Objective-C, an
+    error is raised.
+
+    Supports
+    [Python-style mapping operations](https://docs.python.org/3/library/stdtypes.html#mapping-types-dict)
+    including `__getitem__()`, `__len__()`, `__iter__()`, `__contains__()`, `__eq__()`,
+     `__ne__()`, `copy()`, `get()`, `keys()`, `items()`, and `values()`.
+
+    Unlike most Python mappings, [`NSDictionary`][rubicon.objc.api.NSDictionary]'s
+    [`keys`](https://docs.python.org/3/library/stdtypes.html#dict.keys),
+    [`values`](https://docs.python.org/3/library/stdtypes.html#dict.values), and
+    [`items`](https://docs.python.org/3/library/stdtypes.html#dict.items)
+    methods don't return dynamic views of the dictionary's
+    keys, values, and items.
+
+    [`keys`](https://docs.python.org/3/library/stdtypes.html#dict.keys) and
+    [`values`](https://docs.python.org/3/library/stdtypes.html#dict.values) return
+    lists that are created each time the methods are
+    called, which can have an effect on performance and memory usage for
+    large dictionaries. To avoid this, you can cache the return values of
+    [`keys`](https://docs.python.org/3/library/stdtypes.html#dict.keys) and
+    [`values`](https://docs.python.org/3/library/stdtypes.html#dict.values), or convert
+    the [`NSDictionary`][rubicon.objc.api.NSDictionary] to a Python [`dict`][]
+    beforehand.
+
+    [`items`](https://docs.python.org/3/library/stdtypes.html#dict.items) is currently
+    implemented as a generator, meaning that it returns a single-use iterator. If you
+    need to iterate over
+    [`items`](https://docs.python.org/3/library/stdtypes.html#dict.items) more than
+    once or perform other operations on it, you should convert it to a Python
+    [`set`][] or [`list`][] first.
+    """
     NSMutableDictionary = ObjCClass("NSMutableDictionary")
+    """
+    The
+    [NSMutableDictionary](https://developer.apple.com/documentation/foundation/nsmutabledictionary?language=objc)
+    class from `<Foundation/NSDictionary.h>`.
+
+    This class can be converted to and from Python exactly like its
+    superclass `NSDictionary`.
+
+    Supports
+    [Python-style mutable mapping operations](https://docs.python.org/3/library/stdtypes.html#mapping-types-dict)
+    including `setitem()`, `delitem()`, `clear()`, `pop()`, `popitem()`, `setdefault()`,
+    and `update()`.
+    """
     Protocol = ObjCClass("Protocol")
     """
     The
     [Protocol](https://developer.apple.com/documentation/objectivec/protocol?language=objc)
     class from `<objc/Protocol.h>`.
 
-    /// note | Note
-
     This class has no (non-deprecated) Objective-C methods; protocol objects
     can only be manipulated using Objective-C runtime functions. Rubicon
     automatically wraps all [`Protocol`][rubicon.objc.api.Protocol]
     objects using [`ObjCProtocol`][rubicon.objc.api.ObjCProtocol], which
     provides an easier interface for working with protocols.
-
-    ///
     """
 
 
@@ -1922,9 +2077,9 @@ def py_from_ns(nsobj):
     Currently supported types:
 
     * :class:`~rubicon.objc.runtime.objc_id`: Wrapped in an
-         :class:`ObjCInstance` and converted as below
-    * :class:`NSString`: Converted to :class:`str`
-    * :class:`NSData`: Converted to :class:`bytes`
+         [`ObjCInstance`][rubicon.objc.api.ObjCInstance] and converted as below
+    * :class:`NSString`: Converted to [`str`][]
+    * :class:`NSData`: Converted to [`bytes`][]
     * :class:`NSDecimalNumber`: Converted to :class:`decimal.Decimal`
     * :class:`NSDictionary`: Converted to :class:`dict`, with all keys and
          values converted recursively
@@ -1933,7 +2088,8 @@ def py_from_ns(nsobj):
     * :class:`NSNumber`: Converted to a :class:`bool`, :class:`int` or
          :class:`float` based on the type of its contents
 
-    Other objects are returned unmodified as an :class:`ObjCInstance`.
+    Other objects are returned unmodified as an
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance].
     """
 
     if isinstance(nsobj, (objc_id, Class)):
@@ -1988,11 +2144,11 @@ def ns_from_py(pyobj):
 
     Currently supported types:
 
-    * ``None``, :class:`ObjCInstance`: Returned as-is
+    * ``None``, [`ObjCInstance`][rubicon.objc.api.ObjCInstance]: Returned as-is
     * :class:`enum.Enum`: Replaced by their :attr:`~enum.Enum.value` and
          converted as below
-    * :class:`str`: Converted to :class:`NSString`
-    * :class:`bytes`: Converted to :class:`NSData`
+    * [`str`][]: Converted to :class:`NSString`
+    * [`bytes`][]: Converted to :class:`NSData`
     * :class:`decimal.Decimal`: Converted to :class:`NSDecimalNumber`
     * :class:`dict`: Converted to :class:`NSDictionary`, with all keys and
          values converted recursively
@@ -2001,7 +2157,7 @@ def ns_from_py(pyobj):
     * :class:`bool`, :class:`int`, :class:`float`: Converted to
          :class:`NSNumber`
 
-    Other types cause a :class:`TypeError`.
+    Other types cause a [`TypeError`][].
     """
 
     if isinstance(pyobj, enum.Enum):
@@ -2056,11 +2212,39 @@ if True:
 
 @for_objcclass(Protocol)
 class ObjCProtocol(ObjCInstance):
-    """Python wrapper for an Objective-C protocol."""
+    """Python wrapper for an Objective-C protocol.
+
+    The constructor accepts either the name of an Objective-C protocol to look up
+    (as [`str`][] or [`bytes`][]), or a pointer to an existing protocol object
+    (in any form accepted by [`ObjCInstance`][rubicon.objc.api.ObjCInstance]).
+
+    If given a pointer, it must refer to an Objective-C protocol; pointers
+    to other objects are not accepted. (Use
+    [`ObjCInstance`][rubicon.objc.api.ObjCInstance] to wrap a
+    pointer that might also refer to other kinds of objects.) Creating an
+    :class:`ObjCProtocol` from a ``nil`` pointer returns ``None``.
+
+    :class:`ObjCProtocol` can also be called like :class:`type`, with three
+    arguments (name, bases list, namespace mapping). This form is called
+    implicitly by Python's ``class`` syntax, and is used to create a new
+    Objective-C protocol from Python (see
+    :ref:`custom-classes-and-protocols`). The bases list can contain any
+    number of :class:`ObjCProtocol` objects to be extended by the new
+    protocol.
+
+    If the name of the protocol has already registered with the Objective-C
+    runtime, the ``auto_rename`` option can be used to ensure that the
+    Objective-C name for the new protocol will be unique. A numeric suffix
+    will be appended to the Objective-C name to ensure uniqueness (for
+    example, ``MyProtocol`` will be renamed to ``MyProtocol_2``,
+    ``MyProtocol_3`` etc. until a unique name is found). By default,
+    protocols will *not* be renamed, unless
+    :attr:`ObjCProtocol.auto_rename` is set at the class level.
+    """
 
     @property
     def name(self):
-        """The name of this protocol, as a :class:`str`."""
+        """The name of this protocol, as a [`str`][]."""
 
         return libobjc.protocol_getName(self).decode("utf-8")
 
@@ -2077,32 +2261,32 @@ class ObjCProtocol(ObjCInstance):
     automatically if a protocol with the same name is already exists."""
 
     def __new__(cls, name_or_ptr, bases=None, ns=None, auto_rename=None):
-        """The constructor accepts either the name of an Objective-C protocol to look up
-        (as :class:`str` or :class:`bytes`), or a pointer to an existing protocol object
-        (in any form accepted by :class:`ObjCInstance`).
-
-        If given a pointer, it must refer to an Objective-C protocol; pointers
-        to other objects are not accepted. (Use :class:`ObjCInstance` to wrap a
-        pointer that might also refer to other kinds of objects.) Creating an
-        :class:`ObjCProtocol` from a ``nil`` pointer returns ``None``.
-
-        :class:`ObjCProtocol` can also be called like :class:`type`, with three
-        arguments (name, bases list, namespace mapping). This form is called
-        implicitly by Python's ``class`` syntax, and is used to create a new
-        Objective-C protocol from Python (see
-        :ref:`custom-classes-and-protocols`). The bases list can contain any
-        number of :class:`ObjCProtocol` objects to be extended by the new
-        protocol.
-
-        If the name of the protocol has already registered with the Objective C
-        runtime, the ``auto_rename`` option can be used to ensure that the
-        Objective C name for the new protocol will be unique. A numeric suffix
-        will be appended to the Objective C name to ensure uniqueness (for
-        example, ``MyProtocol`` will be renamed to ``MyProtocol_2``,
-        ``MyProtocol_3`` etc until a unique name is found). By default,
-        protocols will *not* be renamed, unless
-        :attr:`ObjCProtocol.auto_rename` is set at the class level.
-        """
+        # The constructor accepts either the name of an Objective-C protocol to look up
+        # (as [`str`][] or [`bytes`][]), or a pointer to an existing protocol object
+        # (in any form accepted by [`ObjCInstance`][rubicon.objc.api.ObjCInstance]).
+        #
+        # If given a pointer, it must refer to an Objective-C protocol; pointers
+        # to other objects are not accepted. (Use
+        # [`ObjCInstance`][rubicon.objc.api.ObjCInstance] to wrap a
+        # pointer that might also refer to other kinds of objects.) Creating an
+        # :class:`ObjCProtocol` from a ``nil`` pointer returns ``None``.
+        #
+        # :class:`ObjCProtocol` can also be called like :class:`type`, with three
+        # arguments (name, bases list, namespace mapping). This form is called
+        # implicitly by Python's ``class`` syntax, and is used to create a new
+        # Objective-C protocol from Python (see
+        # :ref:`custom-classes-and-protocols`). The bases list can contain any
+        # number of :class:`ObjCProtocol` objects to be extended by the new
+        # protocol.
+        #
+        # If the name of the protocol has already registered with the Objective-C
+        # runtime, the ``auto_rename`` option can be used to ensure that the
+        # Objective-C name for the new protocol will be unique. A numeric suffix
+        # will be appended to the Objective-C name to ensure uniqueness (for
+        # example, ``MyProtocol`` will be renamed to ``MyProtocol_2``,
+        # ``MyProtocol_3`` etc. until a unique name is found). By default,
+        # protocols will *not* be renamed, unless
+        # :attr:`ObjCProtocol.auto_rename` is set at the class level.
 
         if (bases is None) ^ (ns is None):
             raise TypeError("ObjCProtocol arguments 2 and 3 must be given together")
@@ -2176,7 +2360,7 @@ class ObjCProtocol(ObjCInstance):
 
         If the given object is not an Objective-C object, ``False`` is returned.
 
-        This method allows using :class:`ObjCProtocol`\\s as the second argument
+        This method allows using :class:`ObjCProtocol`s as the second argument
         of :func:`isinstance`: ``isinstance(obj, NSCopying)`` is equivalent to
         ``obj.conformsToProtocol(NSCopying)``.
         """
@@ -2190,9 +2374,9 @@ class ObjCProtocol(ObjCInstance):
         """Check whether the given class or protocol conforms to this protocol.
 
         If the given object is not an Objective-C class or protocol,
-        :class:`TypeError` is raised.
+        [`TypeError`][] is raised.
 
-        This method allows using :class:`ObjCProtocol`\\s as the second argument
+        This method allows using :class:`ObjCProtocol`s as the second argument
         of :func:`issubclass`: ``issubclass(cls, NSCopying)`` is equivalent to
         ``cls.conformsToProtocol(NSCopying)``, and ``issubclass(proto,
         NSCopying)`` is equivalent to ``protocol_conformsToProtocol(proto,
@@ -2213,7 +2397,16 @@ class ObjCProtocol(ObjCInstance):
 # Need to use a different name to avoid conflict with the NSObject class.
 # NSObjectProtocol is also the name that Swift uses when importing the NSObject
 # protocol.
-NSObjectProtocol = ObjCProtocol("NSObject")
+# check-docstring-is-first made us do this
+if True:
+    NSObjectProtocol = ObjCProtocol("NSObject")
+    """
+    The
+    [NSObject](https://developer.apple.com/documentation/objectivec/1418956-nsobject?language=objc)
+    protocol from `<objc/NSObject.h>`. The protocol is exported as
+    [`NSObjectProtocol`][rubicon.objc.NSObjectProtocol] in Python because it
+    would otherwise clash with the [`NSObject`][rubicon.objc.NSObject] class.
+    """
 
 
 # When a Python object is assigned to a new ObjCInstance attribute, the Python
@@ -2267,8 +2460,8 @@ except NameError:
 
 
 def objc_const(dll, name):
-    """Create an :class:`ObjCInstance` from a global pointer variable in a
-    :class:`~ctypes.CDLL`.
+    """Create an [`ObjCInstance`][rubicon.objc.api.ObjCInstance] from a global pointer
+    variable in a :class:`~ctypes.CDLL`.
 
     This function is most commonly used to access constant object pointers defined by a
     library/framework, such as
@@ -2368,15 +2561,18 @@ class ObjCBlock:
 
     def __init__(self, pointer, restype=AUTO, *argtypes):
         """The constructor takes a block object, which can be either an
-        :class:`ObjCInstance`, or a raw :class:`~rubicon.objc.runtime.objc_id` pointer.
+        [`ObjCInstance`][rubicon.objc.api.ObjCInstance], or a raw
+        :class:`~rubicon.objc.runtime.objc_id` pointer.
 
-        .. note::
+        /// note | Note
 
-            :class:`~rubicon.objc.runtime.objc_block` is also accepted,
-            because it is a subclass of :class:`~rubicon.objc.runtime.objc_id`).
-            Normally you do not need to make use of this, because in most cases
-            Rubicon will automatically convert
-            :class:`~rubicon.objc.runtime.objc_block`\\s to a callable object.
+        :class:`~rubicon.objc.runtime.objc_block` is also accepted,
+        because it is a subclass of :class:`~rubicon.objc.runtime.objc_id`).
+        Normally you do not need to make use of this, because in most cases
+        Rubicon will automatically convert
+        :class:`~rubicon.objc.runtime.objc_block`s to a callable object.
+
+        ///
 
         In most cases, Rubicon can automatically determine the block's return
         type and parameter types. If a block object doesn't have return/parameter
@@ -2489,10 +2685,12 @@ NOTHING = object()
 class Block:
     """A wrapper that exposes a Python callable object to Objective-C as a block.
 
-    .. note::
+    /// note | Note
 
-        :class:`Block` instances are currently *not* callable from Python,
-        unlike :class:`ObjCBlock`.
+    :class:`Block` instances are currently *not* callable from Python,
+    unlike :class:`ObjCBlock`.
+
+    ///
     """
 
     _keep_alive_blocks_ = {}